pyproject.toml
[project]
name = "vocal-analysis"
version = "0.1.0"
description = "Análise Bioacústica para Classificação de Mecanismos Laríngeos M1/M2 no Choro"
readme = "README.md"
authors = [
    {name = "Arthur Cornélio", email = "arthur.cornelio@gmail.com"}
]
requires-python = ">=3.10"
dependencies = [
    "torchcrepe>=0.0.24",
    "praat-parselmouth>=0.4.3",
    "librosa>=0.10.0",
    "xgboost>=2.0.0",
    "scikit-learn>=1.3.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "torch>=2.0.0",
    "soundfile>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.4.0",
    "jupyter>=1.0.0",
    "ipython>=8.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "ANN", # flake8-annotations
    "B",   # flake8-bugbear
]
ignore = [
    "ANN101", # Missing type annotation for self
    "ANN102", # Missing type annotation for cls
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["ANN"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
